+++ index.mjs
@@ -1,1 +1,1 @@
-export { HamberComponentDev as HamberComponent, afterUpdate, beforeUpdate, createEventDispatcher, getContext, onDestroy, onMount, setContext, tick } from './internal';
+export { HamberComponentDev as HamberComponent, HamberComponentTyped, afterUpdate, beforeUpdate, createEventDispatcher, getAllContexts, getContext, hasContext, onDestroy, onMount, setContext, tick } from './internal/index.mjs';