+++ src/compiler/parse/index.ts
@@ -5,8 +5,15 @@ import { reserved } from '../utils/names';
 import full_char_code_at from '../utils/full_char_code_at';
 import { TemplateNode, Ast, ParserOptions, Fragment, Style, Script } from '../interfaces';
 import error from '../utils/error';
+import parser_errors from './errors';
 
 type ParserState = (parser: Parser) => (ParserState | void);
+
+interface LastAutoClosedTag {
+	tag: string;
+	reason: string;
+	depth: number;
+}
 
 export class Parser {
 	readonly template: string;
@@ -20,6 +27,7 @@ export class Parser {
 	css: Style[] = [];
 	js: Script[] = [];
 	meta_tags = {};
+	last_auto_closed_tag?: LastAutoClosedTag;
 
 	constructor(template: string, options: ParserOptions) {
 		if (typeof template !== 'string') {
@@ -34,7 +42,7 @@ export class Parser {
 			start: null,
 			end: null,
 			type: 'Fragment',
-			children: [],
+			children: []
 		};
 
 		this.stack.push(this.html);
@@ -59,7 +67,7 @@ export class Parser {
 
 		if (state !== fragment) {
 			this.error({
-				code: `unexpected-eof`,
+				code: 'unexpected-eof',
 				message: 'Unexpected end of input'
 			});
 		}
@@ -84,7 +92,7 @@ export class Parser {
 
 	acorn_error(err: any) {
 		this.error({
-			code: `parse-error`,
+			code: 'parse-error',
 			message: err.message.replace(/ \(\d+:\d+\)$/, '')
 		}, err.pos);
 	}
@@ -99,17 +107,18 @@ export class Parser {
 		});
 	}
 
-	eat(str: string, required?: boolean, message?: string) {
+	eat(str: string, required?: boolean, error?: { code: string, message: string }) {
 		if (this.match(str)) {
 			this.index += str.length;
 			return true;
 		}
 
 		if (required) {
-			this.error({
-				code: `unexpected-${this.index === this.template.length ? 'eof' : 'token'}`,
-				message: message || `Expected ${str}`
-			});
+			this.error(error ||
+				(this.index === this.template.length
+					? parser_errors.unexpected_eof_token(str)
+					: parser_errors.unexpected_token(str))
+			);
 		}
 
 		return false;
@@ -162,7 +171,7 @@ export class Parser {
 
 		if (!allow_reserved && reserved.has(identifier)) {
 			this.error({
-				code: `unexpected-reserved-word`,
+				code: 'unexpected-reserved-word',
 				message: `'${identifier}' is a reserved word in JavaScript and cannot be used here`
 			}, start);
 		}
@@ -170,12 +179,13 @@ export class Parser {
 		return identifier;
 	}
 
-	read_until(pattern: RegExp) {
-		if (this.index >= this.template.length)
-			this.error({
-				code: `unexpected-eof`,
+	read_until(pattern: RegExp, error_message?: Parameters<Parser['error']>[0]) {
+		if (this.index >= this.template.length) {
+			this.error(error_message || {
+				code: 'unexpected-eof',
 				message: 'Unexpected end of input'
 			});
+		}
 
 		const start = this.index;
 		const match = pattern.exec(this.template.slice(start));
@@ -192,8 +202,8 @@ export class Parser {
 	require_whitespace() {
 		if (!whitespace.test(this.template[this.index])) {
 			this.error({
-				code: `missing-whitespace`,
-				message: `Expected whitespace`
+				code: 'missing-whitespace',
+				message: 'Expected whitespace'
 			});
 		}
 
@@ -210,27 +220,18 @@ export default function parse(
 	// TODO we may want to allow multiple <style> tags —
 	// one scoped, one global. for now, only allow one
 	if (parser.css.length > 1) {
-		parser.error({
-			code: 'duplicate-style',
-			message: 'You can only have one top-level <style> tag per component'
-		}, parser.css[1].start);
+		parser.error(parser_errors.duplicate_style, parser.css[1].start);
 	}
 
 	const instance_scripts = parser.js.filter(script => script.context === 'default');
 	const module_scripts = parser.js.filter(script => script.context === 'module');
 
 	if (instance_scripts.length > 1) {
-		parser.error({
-			code: `invalid-script`,
-			message: `A component can only have one instance-level <script> element`
-		}, instance_scripts[1].start);
+		parser.error(parser_errors.invalid_script_instance, instance_scripts[1].start);
 	}
 
 	if (module_scripts.length > 1) {
-		parser.error({
-			code: `invalid-script`,
-			message: `A component can only have one <script context="module"> element`
-		}, module_scripts[1].start);
+		parser.error(parser_errors.invalid_script_module, module_scripts[1].start);
 	}
 
 	return {