+++ src/compiler/utils/names.ts
@@ -4,6 +4,7 @@ import full_char_code_at from './full_char_code_at';
 export const globals = new Set([
 	'alert',
 	'Array',
+	'BigInt',
 	'Boolean',
 	'clearInterval',
 	'clearTimeout',
@@ -13,15 +14,19 @@ export const globals = new Set([
 	'decodeURI',
 	'decodeURIComponent',
 	'document',
+	'Element',
 	'encodeURI',
 	'encodeURIComponent',
 	'Error',
 	'EvalError',
 	'Event',
+	'EventSource',
 	'fetch',
+	'FormData',
 	'global',
 	'globalThis',
 	'history',
+	'HTMLElement',
 	'Infinity',
 	'InternalError',
 	'Intl',
@@ -34,6 +39,7 @@ export const globals = new Set([
 	'Math',
 	'NaN',
 	'navigator',
+	'Node',
 	'Number',
 	'Object',
 	'parseFloat',
@@ -49,11 +55,14 @@ export const globals = new Set([
 	'setInterval',
 	'setTimeout',
 	'String',
+	'SVGElement',
+	'Symbol',
 	'SyntaxError',
 	'TypeError',
 	'undefined',
 	'URIError',
 	'URL',
+	'URLSearchParams',
 	'window'
 ]);
 
@@ -105,14 +114,8 @@ export const reserved = new Set([
 	'void',
 	'while',
 	'with',
-	'yield',
+	'yield'
 ]);
-
-const void_element_names = /^(?:area|base|br|col|command|embed|hr|img|input|keygen|link|meta|param|source|track|wbr)$/;
-
-export function is_void(name: string) {
-	return void_element_names.test(name) || name.toLowerCase() === '!doctype';
-}
 
 export function is_valid(str: string): boolean {
 	let i = 0;