+++ src/runtime/internal/utils.ts
@@ -1,3 +1,5 @@
+import { Readable } from 'hamber/store';
+
 export function noop() {}
 
 export const identity = x => x;
@@ -26,7 +28,7 @@ export function blank_object() {
 	return Object.create(null);
 }
 
-export function run_all(fns) {
+export function run_all(fns: Function[]) {
 	fns.forEach(run);
 }
 
@@ -38,8 +40,22 @@ export function safe_not_equal(a, b) {
 	return a != a ? b == b : a !== b || ((a && typeof a === 'object') || typeof a === 'function');
 }
 
+let src_url_equal_anchor;
+
+export function src_url_equal(element_src, url) {
+	if (!src_url_equal_anchor) {
+		src_url_equal_anchor = document.createElement('a');
+	}
+	src_url_equal_anchor.href = url;
+	return element_src === src_url_equal_anchor.href;
+}
+
 export function not_equal(a, b) {
 	return a != a ? b == b : a !== b;
+}
+
+export function is_empty(obj) {
+	return Object.keys(obj).length === 0;
 }
 
 export function validate_store(store, name) {
@@ -56,7 +72,7 @@ export function subscribe(store, ...callbacks) {
 	return unsub.unsubscribe ? () => unsub.unsubscribe() : unsub;
 }
 
-export function get_store_value(store) {
+export function get_store_value<T>(store: Readable<T>): T {
 	let value;
 	subscribe(store, _ => value = _)();
 	return value;
@@ -73,7 +89,7 @@ export function create_slot(definition, ctx, $$scope, fn) {
 	}
 }
 
-export function get_slot_context(definition, ctx, $$scope, fn) {
+function get_slot_context(definition, ctx, $$scope, fn) {
 	return definition[1] && fn
 		? assign($$scope.ctx.slice(), definition[1](fn(ctx)))
 		: $$scope.ctx;
@@ -103,12 +119,28 @@ export function get_slot_changes(definition, $$scope, dirty, fn) {
 	return $$scope.dirty;
 }
 
-export function update_slot(slot, slot_definition, ctx, $$scope, dirty, get_slot_changes_fn, get_slot_context_fn) {
-	const slot_changes = get_slot_changes(slot_definition, $$scope, dirty, get_slot_changes_fn);
+export function update_slot_base(slot, slot_definition, ctx, $$scope, slot_changes, get_slot_context_fn) {
 	if (slot_changes) {
 		const slot_context = get_slot_context(slot_definition, ctx, $$scope, get_slot_context_fn);
 		slot.p(slot_context, slot_changes);
 	}
+}
+
+export function update_slot(slot, slot_definition, ctx, $$scope, dirty, get_slot_changes_fn, get_slot_context_fn) {
+	const slot_changes = get_slot_changes(slot_definition, $$scope, dirty, get_slot_changes_fn);
+	update_slot_base(slot, slot_definition, ctx, $$scope, slot_changes, get_slot_context_fn);
+}
+
+export function get_all_dirty_from_scope($$scope) {
+	if ($$scope.ctx.length > 32) {
+		const dirty = [];
+		const length = $$scope.ctx.length / 32;
+		for (let i = 0; i < length; i++) {
+			dirty[i] = -1;
+		}
+		return dirty;
+	}
+	return -1;
 }
 
 export function exclude_internal_props(props) {
@@ -124,6 +156,14 @@ export function compute_rest_props(props, keys) {
 	return rest;
 }
 
+export function compute_slots(slots) {
+	const result = {};
+	for (const key in slots) {
+		result[key] = true;
+	}
+	return result;
+}
+
 export function once(fn) {
 	let ran = false;
 	return function(this: any, ...args) {
@@ -137,7 +177,7 @@ export function null_to_empty(value) {
 	return value == null ? '' : value;
 }
 
-export function set_store_value(store, ret, value = ret) {
+export function set_store_value(store, ret, value) {
 	store.set(value);
 	return ret;
 }