+++ src/runtime/ssr.ts
@@ -1,0 +1,15 @@
+export {
+	onDestroy,
+	setContext,
+	getContext,
+	getAllContexts,
+	hasContext,
+	tick,
+	createEventDispatcher,
+	HamberComponent,
+	HamberComponentTyped
+} from './index';
+
+export function onMount() {}
+export function beforeUpdate() {}
+export function afterUpdate() {}